Veg votekick commit
This works for day/night/sunset meetings.
Once secondary timer runs out, a veg kick timer starts & is set to 5 minutes.
All alive players are given a kick option. The game will go to the next state once enough players kick, or the veg kick timer runs out. The required number of players to kick is (n / 3) rounded up, where n = number of players alive.
The case where the veg timer has started, and then the vegging player chooses an option, the veg timer is cancelled and the game continues as normal without vegging any players. This works with multiple players too.
Also, the above case will take into account meetings that are noVeg and not noVeg. So if a player has a gun, and hasn't voted, and they are vegging, if they vote, the game will move on to the next state without needing to wait for them to pick a gun option. etc.