Fixed kick shoot bug & fixed no-extend bug
The way the "extending day because of no votes" works, I had to move some code around in the Mafia\Game.js.
For the shooting bug, I added a flag "isVoter" to the Meeting class, and replaced canVote several times in different scripts. This way, isVoter controls whether or not to display the meeting, and canVote allows them to actually vote or not.