Remove birthday purchase & remove year kinda (#540)
* Remove birthday purchase, remove year kinda

I wasn't able to find a way to remove the year detail. I opened a discussion on the repo for the management of the date picker node module. The most I could think to do was to make the minDate = Jan 1st, current year, and then maxDate = Dec 31st, current year.
Secondly, I just removed the limitations on the birthday changes.

* add format

* check bdayChanged to show saveBtn

* fix tabs

---------

Co-authored-by: pikulet <24848927+pikulet@users.noreply.github.com>