Fixed tests
I fixed it so that the game should not create votekicks in the case of a test game.
Also, the masons test was incorrectly done. It only succeeded because the Serial Killer vegged instead of voting no one, since the Serial Killer can't vote no one.
Lastly, I coded the Agent/Spy meetings to be a bit more specific.