Make dark/light theme UX better
The schema is changing from boolean to string, so I tried to compensate for that.
If the "siteColorScheme" setting is set to true in DB, it will display dark theme still.
Likewise, if set to false, it will display light theme.
Honestly it looks like it still worked fine regardless, but put in these checks to make sure.
Also, defaulting to auto setting for best UX.