Merging GY and Village tabs
Added feature that the Graveyard chat and the village chat will be 1 tab.
Also it will work with the party tab.
Styling is such that the avatar will be grayscale, and the user's text will be gray.
Also fixed a bug where going back to a previous night when a party took place would not show the rainbow text.