Combine tabs (#749)
* First progress commit for gy combine

* 2nd progress combine GY tab commit

A little bit closer. Only need 1 part of code to dictate the logic, and still need to change the meeting name. Not sure how to do that, need to think.

* Merging GY and Village tabs

Added feature that the Graveyard chat and the village chat will be 1 tab.
Also it will work with the party tab.
Styling is such that the avatar will be grayscale, and the user's text will be gray.
Also fixed a bug where going back to a previous night when a party took place would not show the rainbow text.

* refactor dead recipient logic to mafia meeting

* refactor dead party logic to mafiaplayer

* clean up the code

* State 0 is usually N0 or day start, include this

* Reworked history to include time of death

---------

Co-authored-by: joyce <24848927+pikulet@users.noreply.github.com>
Co-authored-by: Mino <secondarymino@gmail.com>